updates to KDE. the WM would be great with configuration by text
EVEN JSON. OR A COMMANDLINE INTERFACE LIKE WITH MAC OSX `defaults write`

https://github.com/dcunited001/dc.files.osx/blob/master/osx.sh

figuring out how to remove invisible spacers i accidentally dropped onto the
desktop containment 139 is ridiculous.

AFAI can remember, it's so easy to lose these configurations if they become
corrupted. and without care, this can happen often or on version updates.

not being able to copy text-based definitions for taskbar CPU/GPU/Mem icons
means I have to use screenshots to ensure things are in consistent order or have
consistent configuration. I'm not sure if it's the same source code as Latte,
but these widgets are programmed to avoid problems/bugs. BUT it's the lack of
visibility on your configuration that causes bugs -- no text/json/etc as well as
state mixed with configuration.

like, yeh, it should be user-friendly ... but the KDE/QT configuration interface
needs a major overhaul. I'm not actually sure how Gnome differs, but KDE
dependencies/apps share this and bring it with them to any install they're
on. the only reasonable backup is a total backup of the home
partition/directory, but merging configurations for consistent experiences
between multiple linux installs is a nightmare with that approach.