omfg if i could have just seen it work
the SSH_AUTH_SOCK was never exported for yubikey

... when I tried it two years ago.

so I thought you had to `ssh-add` to the gpg-agent's keystore.
and then i needed to break out all the ssh configs to accomodate
for different keys and in case PIV wasn't available.

never thought to use an alias ... because of course I'm isolated